requirements.txt
python-binance
requests
python-dotenv
tabulate
colorama
pydantic